+++ src/config/config.js
@@ -5,6 +5,29 @@ export const localStorageKeys = {
   redirectLocation: 'redirectLocation',
 }
 
+export const linkPrefix = {
+  article: '/a/',
+  interactiveArticle: '/i/',
+  categories: '/categories/',
+  tag: '/tag/',
+  topic: '/topic/',
+  topics: '/topics/',
+  author: '/author/',
+  authors: '/authors',
+}
+
+export const storage = {
+  google: {
+    schema: 'https',
+    hostname: 'storage.googleapis.com',
+    bucket: 'twreporter-multimedia',
+  },
+}
+
+
+export const mainSite = {
+  url: 'https://www.twreporter.org',
+}
 // export const LOCALSTORAGE_KEY_AUTH = 'authInfo'
 // export const LOCALSTORAGE_KEY_CHECKED = 'checkedInfo'
 // export const LOCALSTORAGE_KEY_REDIRECT_LOCATION = 'redirectLocation'